# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ManageHR API",
        "description": "Documentación de la API RESTful de ManageHR desarrollada en Laravel con autenticación JWT",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/solicitudes-vacaciones-con-archivo": {
            "get": {
                "tags": [
                    "Vacaciones"
                ],
                "summary": "Obtener solicitudes de vacaciones del usuario autenticado",
                "operationId": "d32949f50d0e4656d4177d6fe5f318f3",
                "responses": {
                    "200": {
                        "description": "Lista de solicitudes de vacaciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    },
                    "404": {
                        "description": "No se encontró contrato para el usuario"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vacaciones"
                ],
                "summary": "Registrar una nueva solicitud de vacaciones",
                "operationId": "faa79c614fe8311edd65e005e9bc8c75",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "motivo",
                                    "fechaInicio",
                                    "fechaFinal",
                                    "contratoId"
                                ],
                                "properties": {
                                    "motivo": {
                                        "type": "string",
                                        "example": "Motivo de la solicitud"
                                    },
                                    "fechaInicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-01"
                                    },
                                    "fechaFinal": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-10"
                                    },
                                    "contratoId": {
                                        "type": "integer",
                                        "example": 15
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Solicitud de vacaciones registrada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    },
                    "500": {
                        "description": "Error interno al guardar la solicitud"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/pensiones": {
            "get": {
                "tags": [
                    "Pensiones"
                ],
                "summary": "Listar todas las pensiones",
                "operationId": "35804abd2bfc9d9542de020c63cb7b7b",
                "responses": {
                    "200": {
                        "description": "Lista de pensiones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pensiones": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pensiones/{id}": {
            "patch": {
                "tags": [
                    "Pensiones"
                ],
                "summary": "Intentar actualización parcial (no permitido)",
                "operationId": "b8071c8a5f2cc450d9cdb7d9f97fc898",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la pensión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Actualización no permitida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Introduce tu token JWT en el formato: Bearer {token}"
            }
        }
    },
...
}